Improve dynamic breakpoint handling to support recent runtime changes
