+++ misc/debug-adapter/modules/debug-adapter-core/src/main/java/org/ballerinalang/debugadapter/JBallerinaDebugServer.java
@@ -1013,7 +1013,7 @@ public class JBallerinaDebugServer implements IDebugProtocolServer {
      */
     private void prepareFor(DebugInstruction instruction) {
         clearState();
-        eventProcessor.restoreBreakpoints(instruction);
+        eventProcessor.getBreakpointProcessor().restoreUserBreakpoints(instruction);
         context.setLastInstruction(instruction);
     }
